+++ app/models/user.rb
@@ -46,10 +46,6 @@ class User < ApplicationRecord
     likes.exists?(article_id: article.id)
   end
 
-  def display_name
-    profile&.nickname || self.email.split('@').first
-  end
-
    def follow!(user)
     user_id = get_user_id(user)
 
@@ -66,18 +62,8 @@ class User < ApplicationRecord
     following_relationships.exists?(following_id: user.id)
   end
 
-
-
   def prepare_profile
     profile || build_profile
-  end
-
-  def avatar_image
-    if profile&.avatar&.attached?
-      profile.avatar
-    else
-      'default-avatar.png'
-    end
   end
 
   private
