+++ app/views/blobs/_blob.html.erb
